requirements.txt
asgiref==3.5.2
astroid==2.11.7
autopep8==1.6.0
dill==0.3.5.1
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
Markdown==3.4.1
mccabe==0.7.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
wrapt==1.14.1
